story(development): implements development environment
- adds configuration for development URL
- implement es6 file system

[FINISHES ]
