+++ package.json
@@ -5,8 +5,8 @@
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node app.js",
-    "start:dev": "NODE_ENV=development DEBUG=log nodemon",
+    "start": "node src/app.js",
+    "start:dev": "nodemon --exec babel-node src/app.js",
     "build": "npm run clean && npm run build-server && npm run copy-static-assets"
   },
   "author": "chirchir",
@@ -15,7 +15,15 @@
     "axios": "^0.19.0",
     "body-parser": "^1.18.3",
     "dotenv": "^8.1.0",
-    "express": "^4.16.3",
+    "express": "^4.17.1",
+    "morgan": "^1.9.1",
     "request": "^2.88.0"
+  },
+  "devDependencies": {
+    "@babel/cli": "^7.6.2",
+    "@babel/core": "^7.6.2",
+    "@babel/node": "^7.6.2",
+    "@babel/preset-env": "^7.6.2",
+    "nodemon": "^1.19.2"
   }
 }
